add: mapStyle,sideBarStyle
