+++ src/pages/MainPage/MainPage.jsx
@@ -8,7 +8,7 @@ import SideBar from '../../components/SideBar/SideBar';
 const MainPage = () => {
   return (
     <StMain>
-      {/* <Map /> */}
+      <Map />
       <SideBar>
         <Search />
       </SideBar>
@@ -26,38 +26,4 @@ const StMain = styled.main`
 
 export default MainPage;
 
-const MainContainer = styled.div`
-  display: flex;
-`;
 
-const ModalOverlay = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const ModalContent = styled.div`
-  background: white;
-  padding: 5px;
-  border-radius: 10px;
-  position: relative;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  width: 800px;
-  max-width: 90%;
-`;
-
-const CloseButton = styled.button`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  background: none;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-`;
